+++ ixmp/utils.py
@@ -35,7 +35,7 @@ def as_str_list(arg, idx_names=None):
     Several types of arguments are handled:
     - None: returned as None.
     - str: returned as a length-1 list of str.
-    - list of values: returned as a list with each value converted to str
+    - iterable of values: returned as a list with each value converted to str
     - dict, with list of idx_names: the idx_names are used to look up values
       in the dict, the resulting list has the corresponding values in the same
       order.
@@ -45,6 +45,8 @@ def as_str_list(arg, idx_names=None):
         return None
     elif idx_names is None:
         # arg must be iterable
+        # NB narrower ABC Sequence does not work here; e.g. test_excel_io()
+        #    fails via Scenario.add_set().
         if isinstance(arg, Iterable) and not isinstance(arg, str):
             return list(map(str, arg))
         else:
@@ -438,9 +440,11 @@ def format_scenario_list(platform, model=None, scenario=None, match=None,
 def show_versions(file=sys.stdout):
     """Print information about ixmp and its dependencies to *file*."""
     import importlib
-    from subprocess import DEVNULL, check_output
+    from subprocess import DEVNULL, CalledProcessError, check_output
 
     from xarray.util.print_versions import get_sys_info
+
+    from ixmp.model.gams import gams_version
 
     def _git_log(mod):
         cmd = ['git', 'log', '--oneline', '--no-color', '--decorate', '-n 1']
@@ -490,6 +494,14 @@ def show_versions(file=sys.stdout):
 
         if module_name == 'jpype':
             info.append(('… JVM path', mod.getDefaultJVMPath()))
+
+    # Also display GAMS version, if any
+    try:
+        version = gams_version()
+    except (CalledProcessError, FileNotFoundError):
+        version = "'gams' executable not in PATH"
+    finally:
+        info.extend([("GAMS", version), (None, None)])
 
     # Use xarray to get system & Python information
     info.extend(get_sys_info()[1:])  # Exclude the commit number